# openapi/openapi.yml
openapi: 3.0.0
info:
  title: Simple User API
  description: A simple API to retrieve user information.
  version: 1.0.0
paths:
...
  '/posts/{postId}':
    get:
      description: Retrieve post information by ID.
      parameters:
        - name: postId
          in: path
          description: The unique identifier of the post.
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Post information retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
              examples:
                default:
                  value:
                    id: 1
                    title: 'Hello, World! updated'
                    body: This is my updated post.
                    user_id: 1
                    user_name: John Doe
                    created_at: '2021-01-01T00:00:00Z'
                    updated_at: '2021-01-01T00:00:00Z'
        '404':
          description: Post not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      operationId: getPostById
      summary: ''
      tags:
        - post
    put:
      description: Update a post by ID.
      parameters:
        - name: postId
          in: path
          description: The unique identifier of the post.
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
            examples:
              default:
                value:
                  title: "Hello, World!"
                  body: "This is my first post."
      responses:
        "200":
          description: Post updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePostResponse'
              examples:
                default:
                  value:
                    id: 1
                    title: "Hello, World!"
                    body: "This is my first post."
                    user_id: 1
                    user_name: John Doe
                    created_at: "2021-01-01T00:00:00Z"
                    updated_at: "2021-01-01T00:00:00Z"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      operationId: putPost
      tags:
        - post
      summary: ''
    delete:
      description: Delete a post.
      operationId: deletePost
      parameters:
        - name: postId
          in: path
          description: The unique identifier of the post.
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "204":
          description: no content.
        "400":
          description: Failed to delete a post.
      tags:
        - post
...
components:
  schemas:
...
    Post:
      title: Post
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        user_id:
          type: integer
        user_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - title
        - body
        - user_id
        - user_name
        - created_at
        - updated_at
...
    UpdatePostRequest:
      title: UpdatePostRequest
      type: object
      properties:
        title:
          type: string
        body:
          type: string
      required:
        - title
        - body
    UpdatePostResponse:
      title: UpdatePostResponse
      $ref: '#/components/schemas/Post'
...
servers:
  - url: 'http://localhost:9000'
    description: local server